Google Glass and Pixel 3XL Support
These are the changes necessary to get things running on my particular
setup. I tried to keep the tweaks fairly minimal, the primary goal to
just get things up and running.

Google Glass
============

Google Glass' XE24 update (their final update) uses API 19, so I had
to backport the `minSdkVersion` and make some minor tweaks to call
`getDrawable` through the resource instead of directly.

Additionally, to navigate to the app, a glass specific intent-filter
action had to be added. I believe this should just be ignored in other
contexts, but somebody with a different pair of glasses will need to
test this.

Pixel 3XL (Android 12)
======================

I'm uncertain what the original intent behind the code was in
`isHotspotOn`, but `getDeclaredMethods()` caused the app to crash when
running on my phone. Calling the method directly, though, seemed to
work fine, so I just shifted it to the more direct approach. It seemed
like primarily it was just calling `isWifiApEnabled`, but maybe there
are important use cases I'm bypassing with this approach.